feat: working side lists for editors
