+++ apps/web/src/lib/queries/campaigns.ts
@@ -30,6 +30,12 @@ export const campaignPointsQuery = (
     ] as const,
     queryFn: () => fetchSegmentPoints({ criteria: segmentCriteria, keyFilter }),
     staleTime: Number.POSITIVE_INFINITY,
+    // Release inactive points buffers immediately. Each big campaign is
+    // ~5MB of Float32Array backing memory; accumulating multiple in
+    // cache triggers V8 GC pauses (200-700ms) on later navigations.
+    // Trade-off: revisiting a campaign refetches points (predictable
+    // loading curtain) instead of risking an unexplained freeze.
+    gcTime: 0,
   });
 
 export const campaignKeyCountsQuery = (
