+++ apps/web/src/lib/queries/segments.ts
@@ -31,6 +31,12 @@ export const segmentPreviewQuery = (criteria: Criteria) =>
       return { counts, pointsBuffer };
     },
     staleTime: Number.POSITIVE_INFINITY,
+    // Release inactive points buffers immediately. Each big segment is
+    // ~5MB of Float32Array backing memory; accumulating multiple in
+    // cache triggers V8 GC pauses (200-700ms) on later navigations.
+    // Trade-off: revisiting a segment refetches points (predictable
+    // loading curtain) instead of risking an unexplained freeze.
+    gcTime: 0,
   });
 
 // Binary lng/lat pairs — uploaded directly into a GPU buffer, so the
